pyproject.toml
[project]
name = "iptvportal-client"
version = "0.1.0"
description = "Modern Python client for IPTVPortal JSONSQL API"
authors = [{name = "pv-udpv", email = "pv@example.com"}]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["iptv", "iptvportal", "api", "client", "middleware", "jsonsql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "httpx[http2]>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "orjson>=3.10.0",
    "sqlglot>=27.29.0",
    "typer>=0.20.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-httpx>=0.30.0",
    "pytest-mock>=3.14.0",
    "pytest-timeout>=2.3.1",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
]
cli = [
    "typer[all]>=0.12.0",
    "rich>=13.9.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
iptvportal = "iptvportal.cli.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/iptvportal"]

[dependency-groups]
dev = [
    "ipython>=8.27.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-httpx>=0.35.0",
    "pytest-mock>=3.15.1",
    "pytest-timeout>=2.3.1",
    "ruff>=0.14.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "T10", "RET", "SIM"]
ignore = [
    "N802", "N803", "N806",  # Naming conventions
    "W291", "W293",          # Whitespace issues
    "E712",                  # Comparison to True/False
    "A001", "A002", "A004",  # Shadowing builtins (common in APIs)
    "B904",                  # Exception chaining (can be noisy)
    "DTZ005",                # Timezone-naive datetime (acceptable for logs)
    "E501",                  # Line length (handled by ruff)
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
]
asyncio_mode = "auto"
markers = [
    "asyncio: marks tests as async (using pytest-asyncio)",
    "timeout: per-test timeout in seconds (pytest-timeout)",
]
